remove security server side and leave it to the App
the security algo was duplicated in the express «firewall» and in the App
this might lead to incorrect redirections when the bundle is loading in the browser.

Now it is managed by the App and it use react-router to redirect client side or with SSR
