+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Upload PGP keys using files rather than textboxes.
 
+### Fixed
+
+- double redirections algorithm client and server when not authenticated
+
 ## [0.21] 2019-06-06
 
 ### Added
