+++ src/frontend/web_application/server/app.js
@@ -1,19 +1,20 @@
 import express from 'express';
 import applyAPI from './api';
-import applySecurity from './security';
 import applyAssets from './assets';
 import applyAuth from './auth';
 import applyError from './error';
 import { configure as applyExpressReact } from './express-react';
 import applySSR from './ssr';
+import { httpLog } from './logger';
 
 export default () => {
   const app = express();
 
+  app.use('*', httpLog);
+
   //-------
   // assets always public
   applyAssets(app);
-  applySecurity(app);
   applyAuth(app);
   applyAPI(app);
   applyExpressReact(app);
