+++ src/frontend/web_application/server/auth/lib/seal.js
@@ -4,6 +4,6 @@ export function encode(obj, secret, callback) {
   Iron.seal(obj, secret, Iron.defaults, callback);
 }
 
-export function decode(seal, secret, callback) {
-  Iron.unseal(seal, secret, Iron.defaults, callback);
+export function decode(sealed, secret, callback) {
+  Iron.unseal(sealed, secret, Iron.defaults, callback);
 }
