+++ src/frontend/web_application/server/auth/middlewares/decode-cookie.js
@@ -1,17 +1,18 @@
 import { decode } from '../lib/seal';
+import { COOKIE_NAME } from '../lib/cookie';
 import { getConfig } from '../../config';
 
 export const decodeCookieMiddleware = (req, res, next) => {
   const { seal: { secret } } = getConfig();
-  const seal = req.seal;
+  const cookie = req.signedCookies && req.signedCookies[COOKIE_NAME];
 
-  if (!seal && req.security === false) {
+  if (!cookie) {
     next();
 
     return;
   }
 
-  decode(seal, secret, (err, obj) => {
+  decode(cookie, secret, (err, obj) => {
     if (err || !obj) {
       const error = new Error('Unexpected Server Error on cookie decoding');
       error.status = 500;
@@ -23,6 +24,9 @@ export const decodeCookieMiddleware = (req, res, next) => {
     }
 
     req.user = obj;
+
+    // TODO: refresh token
+    // new Date(Date.UTC(req.user.tokens.expires_at)).getTime() < new Date().getTime() + (1000 * 60 * 10)
 
     next();
   });
