+++ src/frontend/web_application/server/auth/middlewares/index.js
@@ -1,4 +1,1 @@
-export * from './check-cookie';
-export * from './check-token';
 export * from './decode-cookie';
-export * from './catch-login-errors';
