--- src/frontend/web_application/server/security/index.js
@@ -1,5 +1,0 @@
-import { securityMiddleware } from './middlewares';
-
-export default (app) => {
-  app.use(securityMiddleware);
-};
