--- src/frontend/web_application/server/security/middlewares/index.js
@@ -1,1 +1,0 @@
-export * from './security';
