--- src/frontend/web_application/server/security/middlewares/security.spec.js
@@ -1,23 +1,0 @@
-import { securityMiddleware } from './security';
-
-describe('security middleware', () => {
-  it('has an unsecure url', () => {
-    const req = {
-      path: '/auth/signup',
-    };
-
-    securityMiddleware(req, {}, () => {});
-
-    expect(req.security).toEqual(false);
-  });
-
-  it('has a secure url', () => {
-    const req = {
-      path: '/api',
-    };
-
-    securityMiddleware(req, {}, () => {});
-
-    expect(req.security).toEqual(true);
-  });
-});
